requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
tortoise-orm==0.20.0
aerich==0.7.2
aiosqlite==0.17.0

# 任务调度
apscheduler==3.10.4

# 浏览器自动化
playwright==1.41.0

# 内容采集
feedparser==6.0.10
beautifulsoup4==4.12.3
lxml==5.1.0

# LLM提供商
dashscope==1.14.0
openai==1.10.0
qianfan==0.4.0

# 配置
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Agent Skills支持
python-frontmatter==1.0.0

# 工具
aiofiles==23.2.1
python-dateutil==2.8.2
httpx==0.26.0

# 图文渲染
Pillow==10.2.0
Jinja2==3.1.3

# 安全和认证
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
